+++ shardis-api/src/main/java/com/shardis/api/domain/base/BaseEntity.java
@@ -1,16 +1,16 @@
 package com.shardis.api.domain.base;
 
-import lombok.Data;
-
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.MappedSuperclass;
+import java.util.Objects;
+import java.util.UUID;
 
 /**
  * Created by Tomasz Kucharzyk
  */
-@Data
+
 @MappedSuperclass
 public class BaseEntity {
 
@@ -18,4 +18,29 @@ public class BaseEntity {
     @GeneratedValue(strategy = GenerationType.AUTO)
     protected Long id;
 
+    private String uuid = UUID.randomUUID().toString();
+
+
+    public Long getId() {
+        return id;
+    }
+
+    public String getUuid() {
+        return uuid;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+
+        BaseEntity that = (BaseEntity) o;
+
+        return uuid.equals(that.uuid);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(uuid);
+    }
 }