+++ shardis-common/src/main/java/com/shardis/dto/commons/ValueObject.java
@@ -1,13 +1,41 @@
 package com.shardis.dto.commons;
 
-import lombok.AllArgsConstructor;
-import lombok.Data;
 
 /**
  * Created by Tomasz Kucharzyk
  */
-@Data
-@AllArgsConstructor
 public class ValueObject {
-    String value;
+
+    private String value;
+
+    public ValueObject() {
+    }
+
+    public ValueObject(String value) {
+        this.value = value;
+    }
+
+    public String getValue() {
+        return value;
+    }
+
+    public void setValue(String value) {
+        this.value = value;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+
+        ValueObject that = (ValueObject) o;
+
+        return value != null ? value.equals(that.value) : that.value == null;
+
+    }
+
+    @Override
+    public int hashCode() {
+        return value != null ? value.hashCode() : 0;
+    }
 }