+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 * Updated Angular to rc.4
 * Switched back to awesome-ts-loader (faster)
 * Updated Spring Cloud to Brixton.SR1
+* Removed lombok from project
+* Added UUID to all entities
 
 ### 1.8.0 (30.06.2016)
 * Better directory structure for angular components